backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
mysqlclient==2.2.0
marshmallow==3.19.0
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0 